Exclude external source files from swiftformat
